+++ R/utils-julia.R
@@ -1,38 +1,14 @@
-is_jl <- function(x, type) {
-  inherits(x, "JuliaProxy") &&
-    if (!missing(type)) { type %in% jl_supertypes(x) } else { TRUE }
-}
-
-jl_get <- function(x) {
-  if (is_jl(x)) {
-    x <- JuliaConnectoR::juliaGet(x)
-    JL_attr <- grep(x = names(attributes(x)), "^JL[A-Z]+$", value = TRUE)
-    attributes(x)[JL_attr] <- NULL
-  }
-  x
-}
-
-jl_evalf <- function(x, ...) {
-  if (is.null(x)) return(NULL)
-  dots <- list(...)
-  if (length(dots) == 0) {
-    JuliaConnectoR::juliaEval(x)
-  } else {
-    JuliaConnectoR::juliaEval(sprintf(x, ...))
-  }
-}
-
 jl_io <- function(verbose) {
   if (verbose) "" else "io=devnull"
 }
 
 jl_pkg_installed <- function(x, ..., verbose = interactive()) {
-  jl_evalf('!isnothing(Pkg.status("%1$s"; %2$s))', x, jl_io(verbose))
+  jl('!isnothing(Pkg.status("%1$s"; %2$s))', x, jl_io(verbose), .R = TRUE)
 }
 
 jl_pkg_add <- function(x, ..., verbose = interactive()) {
   if (!jl_pkg_installed(x, verbose = verbose)) {
-    jl_evalf('Pkg.add("%2$s"; %1$s); using %2$s', jl_io(verbose), x)
+    jl('Pkg.add("%2$s"; %1$s); using %2$s', jl_io(verbose), x)
   }
 }
 
@@ -62,6 +38,10 @@ jl_supertypes <- function(x) {
         supertypes
     end
   ", x = x)
-  vec <- unlist(JuliaConnectoR::juliaGet(supertypes))
+  vec <- unlist(jl_get(supertypes))
   gsub("\\{.*\\}$", "", vec)
+}
+
+jl_format <- function(x, ...) {
+  JuliaConnectoR::juliaCall("JuliaFormatter.format_text", x, align_matrix = TRUE, ...)
 }